pyproject.toml
[project]
name = "corp-assistant"
version = "0.1.0"
description = ""
authors = [
    {name = "mangtaeeee",email = "osp9658@naver.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "langgraph (>=0.3.24,<0.4.0)",
    "langchain-openai (>=0.3.12,<0.4.0)",
    "langchain-community (>=0.3.20,<0.4.0)",
    "qdrant-client (>=1.13.3,<2.0.0)",
    "pydantic (>=2.11.1,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-qdrant (>=0.2.0,<0.3.0)"
]

[tool.poetry]
packages = [{include = "corp_assistant", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
